fix postgres int types to bigint

+++ internal/domain/car.go
@@ -12,8 +12,8 @@ type Car struct {
 	CarId      uint64   `json:"car_id" omitempty:"true"`
 	Mark       string   `json:"mark" required:"true"`
 	Model      string   `json:"model" required:"true"`
-	OwnerCount uint     `json:"owner_count"`
-	Price      uint     `json:"price"`
+	OwnerCount uint64   `json:"owner_count"`
+	Price      uint64   `json:"price"`
 	Currency   Currency `json:"currency"`
 	Options    []string `json:"options" required:"true"`
 }
